update :: size, postSequence to int

+++ bitgouel-api/src/main/kotlin/team/msg/domain/post/presentation/web/QueryAllPostsWebRequest.kt
@@ -7,7 +7,7 @@ data class QueryAllPostsWebRequest(
     @RequestParam(required = false)
     val postSequence: Int?,
     @RequestParam
-    val size: Long,
+    val size: Int,
     @RequestParam(required = false)
     val type: FeedType?
 )
